code refactor and add comment
