+++ app/src/config/common.js
@@ -11,7 +11,8 @@ export const STYLE = {
 export const LINK = {
     FACEBOOK: 'https://www.facebook.com/over.party.lab/',
     INSTAGRAM: 'https://www.instagram.com/over.party.lab/',
-    MEDIUM: 'https://medium.com/@overpartylab'
+    MEDIUM: 'https://medium.com/@overpartylab',
+    GITHUB: 'https://github.com/overpartylab/cocktails-guide-book'
 };
 
 export const ROUTE = {
