--- app/src/config/contact.js
@@ -1,6 +1,0 @@
-export const LINK = {
-    FACEBOOK: 'https://www.facebook.com/over.party.lab/',
-    INSTAGRAM: 'https://www.instagram.com/over.party.lab/',
-    MEDIUM: 'https://medium.com/@overpartylab',
-    GITHUB: 'https://github.com/overpartylab/cocktails-guide-book'
-};
