+++ app/src/container/header/sideNav.js
@@ -1,13 +1,16 @@
+// core
 import React from 'react';
 import { Link } from "react-router-dom";
 import PropTypes from 'prop-types';
 
+// third party component
 import Drawer from '@material-ui/core/Drawer';
 import IconButton from '@material-ui/core/IconButton';
 import SearchTwoToneIcon from '@material-ui/icons/SearchTwoTone';
 import ContactMailTwoToneIcon from '@material-ui/icons/ContactMailTwoTone';
 import ChevronRightIcon from '@material-ui/icons/ChevronRight';
 
+// components
 import { LogoImage } from '../style.css.js';
 import {
     SideNavContainer,
@@ -19,8 +22,9 @@ import {
     SideNavFooter
 } from './sideNav.css.js';
 
-import COMMON_TEXT from '../../assets/wording/common.json';
+// utils, config and assets
 import { ROUTE, COMMON } from '../../config/common';
+import COMMON_TEXT from '../../assets/wording/common.json';
 import CONTACT_TEXT from '../../assets/wording/contact.json';
 
 const SideNav = ({ open, onClose }) => {
